requirements.txt
boto3
ffmpeg
ffmpeg-python
imutils
opencv-python
pillow
facenet-pytorch
numpy
torch
torchvision

